backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
ultralytics==8.3.251
pillow>=11.0.0
pyyaml==6.0.1
aiofiles==23.2.1
psutil==5.9.8
# PyTorch with CUDA 12.8 support for RTX 5070 (Blackwell sm_120)
torch>=2.0.0
torchvision>=0.15.0
